Fixing \n newline add to string and textbox.
The copy past process with the clipboard was adding extra new to the textbox showing the converted text and also when the string was copied to clipboard.  I removed the end characters to take away the added \n newlines.  And when textbox result is copied to clipboard I removed that from the string.  These apparently are added by default.  I have not tested in Linux yet, only Windows